LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-none-block-expected.txt

CONSOLE MESSAGE: Refused to display 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=%27none%27' in a frame because an ancestor violates the following Content Security Policy directive: "frame-ancestors 'none'".
A 'frame-ancestors' CSP directive with a value 'none' should block rendering.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


IFrame load event fired: the IFrame is cross-origin (or was blocked).
PASS The IFrame should have been blocked (or cross-origin). It was.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
